Move other ASSERT_EQUAL_XXX macros to a header too
Prepare for reusing them from other source files.

Note that this requires using test_context from test-assert.h, which now
has to include test-context.h.
